+++ src/win32/view.c
@@ -294,6 +294,7 @@ ctd_status ctd_view_measure(ctd_handle widget, double avail_width, double avail_
             if (height < 17.0) height = 17.0;
             break;
         case CTD_W_TEXT_FIELD:
+        case CTD_W_SECURE_FIELD:
             width += 8.0;
             height += 8.0;
             break;